requirements.txt
streamlit
requests
openai
pandas
beautifulsoup4
python-dotenv
